requirements.txt
streamlit>=1.28.0
numpy>=1.24.0
scipy>=1.10.0
librosa>=0.10.0
soundfile>=0.12.0
matplotlib>=3.7.0
plotly>=5.14.0
anthropic>=0.18.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
noisereduce>=3.0.0
transformers>=4.30.0
torch>=2.0.0
torchaudio>=2.0.0
psutil>=5.9.0
asyncio>=3.4.3 